afc-monitor-cli/requirements.txt
art==6.2
click==8.1.7
python-dotenv==1.0.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
dependency-injector==4.41.0